Continued to fix memory allocation/addressing issues with staggered vs nonstaggered arrays. Testing hasn't been super thorough, but noticed several arrays were padded with zeroes in weird places due to the way MPI_Gather works. I think this has been fixed, as now the vector and scalar arrays index properly without weird zeroes being present. I guess we'll see if its actually correct soon.

+++ datastructs.cu
@@ -220,7 +220,7 @@ void deallocate_parcels_cpu(parcel_pos *parcels) {
    fields for integration and calculation. This
    only ever gets called by Rank 0, so there 
    should be no need for a CPU counterpart. */
-integration_data* allocate_integration_managed(long bufsize) {
+integration_data* allocate_integration_managed(int bufsize) {
     integration_data *data;
     // create the struct on both the GPU and the CPU.
     cudaMallocManaged(&data, sizeof(integration_data));
